Добавил интерфейсы репозиториев, их реализацию. Правки в репозиториях, и реализациях.

+++ src/main/java/ru/job4j/cars/repository/file/FileRepository.java
@@ -9,9 +9,9 @@ public interface FileRepository {
 
     File save(File file);
 
-    Optional<File> findById(int id);
+    Optional<File> findById(Long id);
 
     List<File> findAll();
 
-    boolean deleteById(int id);
+    boolean deleteById(Long id);
 }
