+++ src/main/java/ru/job4j/cars/repository/file/HQLFileRepository.java
@@ -22,7 +22,7 @@ public class HQLFileRepository implements FileRepository {
     }
 
     @Override
-    public Optional<File> findById(int id) {
+    public Optional<File> findById(Long id) {
         return crudRepository.optional(
                 "FROM File WHERE id = :fId",
                 File.class,
@@ -37,7 +37,7 @@ public class HQLFileRepository implements FileRepository {
     }
 
     @Override
-    public boolean deleteById(int id) {
+    public boolean deleteById(Long id) {
         return crudRepository.runBoolean("DELETE File WHERE id = :fId",
                 Map.of("fId", id)
         );
